ath10k: report tx-rate when using CT firmware.
This takes the tx-rate-code and tx-rate-flags reported
by the CT firmware and decodes it.  I suspect the index
might need special mapping from firmware's idea of the
rate index to the linux's idea.  But, it seems to at
least partially work as is.

Backport feature flags into this commit because we need
added checks to work with modern wave-2 firmware.

Signed-off-by: Ben Greear <greearb@candelatech.com>