ath9k:
Having a small number of BSS (compared to ATH_BCBUF) has the disadvantage,
that stations would be awake longer than necessary due to the buffered
packets being served round-robin instead of first all of one bss, and
then all of the other, with stations of either bss sleeping when the
other bss talking.

To overcome this issue, this patch assigns the slots in a way that maximizes
the distance - so that if there aren't too many buffered packets, only the
right stations would be awake.

Original Patch is by M. Braun:  michael-dev <michael-dev@fami-braun.de>

Signed-off-by: Ben Greear <greearb@candelatech.com>