mtk: mac80211: legacy AP scan request should contain valid channels
In single-wiphy, if scan_freqs is not specified in scan request,
mac80211 will trigger a scan that includes all bands.
However, legacy AP should only scan the channels of its operating band.

This commit adds the checks for including valid channels in legacy AP
scan.

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>

preset_chandef is only set on the first wdev of each phy, so we refer to
another data structure for the band of current wdev.

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>