mtk: mac80211: fixup sanity check in ieee80211_build_hdr()
This might be a typo from upstream.

There were two conditions in the if-clause:
1. pre_conf_link_id != link_id
2. link_id != IEEE80211_LINK_UNSPECIFIED
and if the two conditions are true at the same time, the frame is
dropped.

The link_id is originally set to pre_conf_link_id and might be changed
to sta->deflink.link_id if sdata is an AP interface. In other word, the
only case that link_id is not equal to pre_conf_link is link_id becomes
sta->deflink.link_id, which is nerver IEEE80211_LINK_UNSPECIFIED.
In sammary, if the condition 1 is satisfied, the condition 2 must also
be satisfied and therefore becomes redundant.

The condition 2 might be a check on pre_conf_link_id not being
IEEE80211_LINK_UNSPECIFIED to prevent dropping the frame when then
link id is not specified originally.

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>