mtk: mt76: mt7996: Remove mutex in atomic operation
This commit introduces two changes. First, do not hold the mutex in the
atomic operation function, as it may cause a call trace when enabling
the kernel build option "Compile the kernel with sleep inside atomic
section checking." This mutex is removed since there is no significant
benefit to use it. Second, iterate over the link station instead of the
stations to find the minimum RSSI for each band.

Signed-off-by: Howard Hsu <howard-yh.hsu@mediatek.com>