wifi: mt76: mt7996: Refactor of rx filter logic.
The existing logic for configuring the rx filter is hard to follow and
contained at least one sneaky bug: since mt7996_set_monitor(_, false) is
only called when a monitor is removed, MT_WF_RFCR_DROP_OTHER_UC would
not be configured on a fresh boot. This meant that the rxfilter was too
verbose until a monitor had been added and removed. This refactor hopes
to avoid sequence-of-action bugs like this by:

1. Placing all logic that directly alters the rx filter in one function,
   so that state does not need to be tracked across methods.
2. Explicitly tracking all values sent to the control registers within
   the radio phy, for easier debugging.
3. Realizing the rx filter control registers freshly every time that
   mt7996_phy_set_rxfilter() is called.

In order to perform goals (1) and (3), new state needed to be added
to the radio phy to remember the mac80211 rxfilter flags that had been
requested, and to track if the radio is acting as a monitor.

Since the existing MT76_FILTER macro could not handle all cases where
the rxfilter needs to be configured, it was not used here.

Fixes: 69d54ce7491d ("wifi: mt76: mt7996: switch to single multi-radio wiphy")
Signed-off-by: Rory Little <rory@candelatech.com>