wifi: mt76: mt7996: Add bss_idx field to mt76_vif_link
Repeater BSSID links share a BSS index on each band. To minimize the
changes needed, we only want to change the indices for these repeater
links, therefore these special shared indices cannot start at zero, and
instead need to sit after every other BSS index. However, this is an
issue in generic mt76 code, where the number of valid links can vary
across devices.

The solution to this is to introduce a bss_idx field, which when zero
indicates that a normal indexing scheme, based on the VIF index, should
be used. However, when nonzero, it indicates that a custom BSS index,
equal to one less than the indicated value, should be used. This can be
configured by links which require a different BSS index, such as those
using repeater BSSIDs.

Signed-off-by: Rory Little <rory@candelatech.com>