wifi: mt76: mt7996: Add bss_idx field to mt76_vif_link
There are situations (such as using repeater stations) when the BSS
index may not match the VIF index. In these cases, we need to keep track
of a separate BSS index. To avoid code churn on existing drivers, the
value of 0 for this index indicates that the VIF index should be used.
Therefore, the bss_idx field is one higher than its intended value, and
should be adjusted whenever placed into an MCU message.

Signed-off-by: Rory Little <rory@candelatech.com>