mac80211:  Fix scan regression for be200 radios.
Recent mtk/felix patches from mtk owrt tree broke scanning
for drivers that do not set n_radio > 0.  Assume that if it is
zero, then there is one radio.  This allows scanning to
work for be200, though perhaps it is too simplistic of a fix.

Add logging around scan failures.

Signed-off-by: Ben Greear <greearb@candelatech.com>