wifi: mt76: mt7915: Add vif-level set_rate_override debugfs file
This is tied as a feature to VIFs, fundamentally. Therefore, it should
exist at a VIF level. This saves us from needing to pass a VIF to the
file for lookup.

Also fix a bug for AP VIFs, where the testmode config was not being
correctly propogated to new stations.

Signed-off-by: Rory Little <rory@candelatech.com>