wifi: mt76: mt7996: Re-index STA setup TLV bss_index.
This caused issues with MLD links. Links likely need a consistent index
across all TLVs, not just within the setup TLV. Bump repeater index to
the end of the block, instead of the front, so that everything else is
unaffected by the shift.

We still keep the bss_idx field, for now. This is because the base index
to collapse repeater BSSIDs onto could vary from driver to driver, and
so this is a way to allow specific drivers to indicate that base. In
the future, it may just be better to put this base index onto the phy,
for more generic use.

Signed-off-by: Rory Little <rory@candelatech.com>