wifi: mac80211: Prevent multiple active links for MLSR-only devices.
The first time set_vif_links_bitmap is called, it configures just the
association link, which is good enough for MLSR devices. However, for
MLMR, we would like to be able to come in a second time to configure
more than one active link. This incompatibility of behavior corresponds
well with the WANTS_VALID_LINKS hardware flag, so use that to determine
whether or not to allow configuration of multiple links.

Fixes: ("mtk: mac80211: allow multiple links for STA vif")
Signed-off-by: Rory Little <rory@candelatech.com>