wifi: mt76: mt7996: Do not exit vif_link_add between wcid alloc and init
The WCID is not in a valid state until it has been fully initialized,
and cleanup code in vif_link_remove can only see if the WCID has been
allocated or not. Therefore, we need to be careful to not go to cleanup
logic while the WCID is allocated, but in an uninitialized state.

Signed-off-by: Rory Little <rory@candelatech.com>