wifi: Support disabling 40 and 80Mhz
Check assoc flags to see if we are requested to disable 40 or 80mhz
bandwidths.

Rename flags to be more descriptive of their action.  Should not change
binary API with older user-space.

This lets me assoc a EHT STA at 40Mhz, but peer AP is confused somehow,
and does not seem to attempt 40Mhz tx of frames, so more work is needed.

Signed-off-by: Ben Greear <greearb@candelatech.com>