wifi: mt76: mt7996: guard NPE in debugfs link info
When querying the `mt76_links_info` debugfs file at specific time, a
kernel NPE can be hit. There's a brief moment when a station is just created
before connecting or after deauthenticating where the channel context is null,
but the device is still alive. Fix is to simply expand the existing
similar guard.

Signed-off-by: Dylan Eskew <dylan.eskew@candelatech.com>